# BurlywoodWhimsicalApplets/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MarketplacePrice {
  marketplace: string;
  product: string;
  price: number;
  difference: number;
  confidence: number;
  status: string;
}

export type ProductPrices = {[key: string]: number};

export interface HistoryPoint {
  label: string;
  gem: number;
  amazon: number;
  flipkart: number;
  indiamart: number;
}

export interface Product {
  id: string;
  name: string;
  brand: string;
  model: string;
  category: string;
  gemPrice: number;
  prices: ProductPrices;
  confidence: number;
  history: HistoryPoint[];
}

export interface Comparison {
  product: Product;
  rows: MarketplacePrice[];
  marketAverage: number;
  lowestMarketPrice: number;
  highestMarketPrice: number;
  difference: number;
  percentageDifference: number;
  potentialSavings: number;
  recommendation: string;
}

export interface PriceHistory {
  product: Product;
  points: HistoryPoint[];
  currentGemPrice: number;
  marketAverage: number;
  change30d: number;
  change90d: number;
}

export type DashboardReviewItemsItem = {
  name: string;
  difference: number;
};

export type DashboardCategoryBreakdownItem = {
  category: string;
  count: number;
};

export interface Dashboard {
  productsAnalyzed: number;
  aboveMarket: number;
  belowMarket: number;
  potentialSavings: number;
  reviewItems: DashboardReviewItemsItem[];
  categoryBreakdown: DashboardCategoryBreakdownItem[];
}

export interface RefreshResult {
  message: string;
  lastUpdated: string;
}

export type ListProductsParams = {
category?: string;
q?: string;
};