# BurlywoodWhimsicalApplets/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Cached product comparison data
  - name: analytics
    description: Buyer analytics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List cached products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
  /compare/{id}:
    get:
      operationId: compareProduct
      tags: [products]
      summary: Compare a product across marketplaces
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comparison"
        "404":
          description: Product not found
  /history/{id}:
    get:
      operationId: getPriceHistory
      tags: [products]
      summary: Get price history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistory"
        "404":
          description: Product not found
  /dashboard:
    get:
      operationId: getDashboard
      tags: [analytics]
      summary: Get buyer dashboard metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /refresh:
    post:
      operationId: refreshPrices
      tags: [products]
      summary: Refresh cached prices
      responses:
        "200":
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResult"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MarketplacePrice:
      type: object
      required: [marketplace, product, price, difference, confidence, status]
      properties:
        marketplace: { type: string }
        product: { type: string }
        price: { type: number }
        difference: { type: number }
        confidence: { type: number }
        status: { type: string }
    Product:
      type: object
      required: [id, name, brand, model, category, gemPrice, prices, confidence, history]
      properties:
        id: { type: string }
        name: { type: string }
        brand: { type: string }
        model: { type: string }
        category: { type: string }
        gemPrice: { type: number }
        prices:
          type: object
          additionalProperties: { type: number }
        confidence: { type: number }
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryPoint"
    Comparison:
      type: object
      required: [product, rows, marketAverage, lowestMarketPrice, highestMarketPrice, difference, percentageDifference, potentialSavings, recommendation]
      properties:
        product: { $ref: "#/components/schemas/Product" }
        rows:
          type: array
          items: { $ref: "#/components/schemas/MarketplacePrice" }
        marketAverage: { type: number }
        lowestMarketPrice: { type: number }
        highestMarketPrice: { type: number }
        difference: { type: number }
        percentageDifference: { type: number }
        potentialSavings: { type: number }
        recommendation: { type: string }
    HistoryPoint:
      type: object
      required: [label, gem, amazon, flipkart, indiamart]
      properties:
        label: { type: string }
        gem: { type: number }
        amazon: { type: number }
        flipkart: { type: number }
        indiamart: { type: number }
    PriceHistory:
      type: object
      required: [product, points, currentGemPrice, marketAverage, change30d, change90d]
      properties:
        product: { $ref: "#/components/schemas/Product" }
        points:
          type: array
          items: { $ref: "#/components/schemas/HistoryPoint" }
        currentGemPrice: { type: number }
        marketAverage: { type: number }
        change30d: { type: number }
        change90d: { type: number }
    Dashboard:
      type: object
      required: [productsAnalyzed, aboveMarket, belowMarket, potentialSavings, reviewItems, categoryBreakdown]
      properties:
        productsAnalyzed: { type: number }
        aboveMarket: { type: number }
        belowMarket: { type: number }
        potentialSavings: { type: number }
        reviewItems:
          type: array
          items:
            type: object
            required: [name, difference]
            properties:
              name: { type: string }
              difference: { type: number }
        categoryBreakdown:
          type: array
          items:
            type: object
            required: [category, count]
            properties:
              category: { type: string }
              count: { type: number }
    RefreshResult:
      type: object
      required: [message, lastUpdated]
      properties:
        message: { type: string }
        lastUpdated: { type: string }