# BurlywoodWhimsicalApplets/lib/api-zod/src/generated/types/comparison.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketplacePrice } from './marketplacePrice';
import type { Product } from './product';

export interface Comparison {
  product: Product;
  rows: MarketplacePrice[];
  marketAverage: number;
  lowestMarketPrice: number;
  highestMarketPrice: number;
  difference: number;
  percentageDifference: number;
  potentialSavings: number;
  recommendation: string;
}