# BurlywoodWhimsicalApplets/lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardCategoryBreakdownItem } from './dashboardCategoryBreakdownItem';
import type { DashboardReviewItemsItem } from './dashboardReviewItemsItem';

export interface Dashboard {
  productsAnalyzed: number;
  aboveMarket: number;
  belowMarket: number;
  potentialSavings: number;
  reviewItems: DashboardReviewItemsItem[];
  categoryBreakdown: DashboardCategoryBreakdownItem[];
}