# BurlywoodWhimsicalApplets/lib/api-zod/src/generated/types/historyPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryPoint {
  label: string;
  gem: number;
  amazon: number;
  flipkart: number;
  indiamart: number;
}