# BurlywoodWhimsicalApplets/lib/api-zod/src/generated/types/marketplacePrice.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplacePrice {
  marketplace: string;
  product: string;
  price: number;
  difference: number;
  confidence: number;
  status: string;
}