# BurlywoodWhimsicalApplets/lib/api-zod/src/generated/types/priceHistory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryPoint } from './historyPoint';
import type { Product } from './product';

export interface PriceHistory {
  product: Product;
  points: HistoryPoint[];
  currentGemPrice: number;
  marketAverage: number;
  change30d: number;
  change90d: number;
}