# BurlywoodWhimsicalApplets/lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryPoint } from './historyPoint';
import type { ProductPrices } from './productPrices';

export interface Product {
  id: string;
  name: string;
  brand: string;
  model: string;
  category: string;
  gemPrice: number;
  prices: ProductPrices;
  confidence: number;
  history: HistoryPoint[];
}